requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
distro==1.9.0
Django==4.2.14
django-activity-stream==2.0.0
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.35.13
phonenumbers==8.13.40
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
whitenoise==6.7.0
